+++ src/MessageService.php
@@ -12,8 +12,6 @@ namespace LitGroup\Sms;
 
 use LitGroup\Sms\Exception\GatewayException;
 use LitGroup\Sms\Gateway\GatewayInterface;
-use LitGroup\Sms\Logger\MessageLoggerInterface;
-use LitGroup\Sms\Logger\NullMessageLogger;
 use Psr\Log\LoggerAwareInterface;
 use Psr\Log\LoggerInterface;
 use Psr\Log\NullLogger;
@@ -31,11 +29,6 @@ class MessageService implements MessageServiceInterface, LoggerAwareInterface
     private $gateway;
 
     /**
-     * @var MessageLoggerInterface
-     */
-    private $messageLogger;
-
-    /**
      * @var LoggerInterface
      */
     private $logger;
@@ -47,7 +40,6 @@ class MessageService implements MessageServiceInterface, LoggerAwareInterface
     public function __construct(GatewayInterface $gateway)
     {
         $this->gateway = $gateway;
-        $this->messageLogger = new NullMessageLogger();
         $this->logger = new NullLogger();
     }
 
@@ -58,7 +50,6 @@ class MessageService implements MessageServiceInterface, LoggerAwareInterface
     {
         try {
             $this->gateway->sendMessage($message);
-            $this->messageLogger->addMessage($message);
 
             $this->logger->info('Message (SMS) was sent.', [
                 'message' => [
@@ -74,14 +65,6 @@ class MessageService implements MessageServiceInterface, LoggerAwareInterface
 
             throw $e;
         }
-    }
-
-    /**
-     * @inheritDoc
-     */
-    public function setMessageLogger(MessageLoggerInterface $messageLogger)
-    {
-        $this->messageLogger = $messageLogger;
     }
 
     /**